# src/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Product": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"stockQuantity": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"price",
					"stockQuantity",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProductDetails": {
				"properties": {
					"name": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"stockQuantity": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"name",
					"price",
					"stockQuantity"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "exercise",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/products/{id}": {
			"get": {
				"operationId": "GetProductById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Product"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "Product not found"
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Product"
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ProductDetails"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteProduct",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/products": {
			"get": {
				"operationId": "GetAllProducts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Product"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Product"
								}
							}
						}
					},
					"201": {
						"description": "Product created"
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ProductDetails"
							}
						}
					}
				}
			}
		},
		"/products/name/{name}": {
			"get": {
				"operationId": "GetProductsByName",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Product"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "name",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/products/count/all": {
			"get": {
				"operationId": "CountProducts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"format": "double"
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}